Fix issue with simultaneous CQI and ACK/NACK transmission in CA (#1067)
* Fix memory corruption when phy calling mac scheduler and not yet initiated

* Do not drop CQI if collision with ACK/NACK and PUSCH

* Allocate CQI resources for SCell properly

* Use UE_PCELL_CC_IDX macro

* Protect ul_sched from being called if not yet started

+++ lib/include/srslte/phy/phch/cqi.h
@@ -55,7 +55,7 @@ typedef enum {
 typedef struct {
   bool                     periodic_configured;
   bool                     aperiodic_configured;
-  uint32_t                 pmi_idx;
+  uint16_t                 pmi_idx;
   uint32_t                 ri_idx;
   bool                     ri_idx_present;
   bool                     format_is_subband;
